backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
Pillow
pyzbar
opencv-python-headless
pytesseract
playwright